Fix tormented audio scheduling bug
In audio worklets, the process method is called synchronously. However,
I've gone with the bright idea of using async/await for music
sequencing. I still think it's kinda neat, but it also summoned a
demonic bug.

What's happening is that the JS microtask queue (containing things like
continuations of async functions) only gets executed after the process
method, therefore at the end of a block containing 128 audio frames.

So when you call `await sequence.play(2)`, you're not really playing 2
seconds, you're playing 2 seconds plus whatever remaining samples are
left in the audio block.

This is typically a very small error, but some patches are playing
tons of small intervals, and the timing error accumulates. This became
noticeable when the CLI and web renderer sounded different, because
they yielded to the JS microtask queue at different points in the audio
stream.

Here I implement one part of a 2-pronged solution. This part is the
compensation of timing error when calling sequence.play. The other part
is ensuring that the web and CLI versions give control to the microtask
queue at the exact same intervals, this is done in the Teasynth
project. Is it a hack? Yes!